Top viewport renders using a .obj for face vertices
The view won't render anything if the 3 colors are not chosen from the
palette tables, so this will be addressed to allow the default color be
a transluscent grey.

Added a info_panel for basic debugging information live while in the
app.

Need to create a file dialog custom subclass for a modal file opening in
one batch